Added new built-in types
Early support for some the proposed types using templates
Fixed some compiler warnings regarding the double to float conversion
Added a hook to return a table from the C api

+++ src/luaexpose/LuaContextBase.cpp
@@ -19,7 +19,7 @@ float LuaContextBase::getNumberFromStack( int stackPos )
 {
 	// Get a number from a stack position
 
-	return( lua_tonumber( m_L, stackPos ) );
+	return( static_cast<float>( lua_tonumber( m_L, stackPos ) ) );
 }
 
 int LuaContextBase::getIntegerFromStack( int stackPos )
